requirements.txt
discord.py==1.7.2
PyNaCl==1.4.0
rich==10.2.2